chore(web): also splits BannerView and BannerController into separate files
