+++ web/src/engine/osk/src/index.ts
@@ -4,7 +4,7 @@ export { default as OSKView } from './views/oskView.js';
 export { default as FloatingOSKView, FloatingOSKViewConfiguration } from './views/floatingOskView.js';
 export { default as AnchoredOSKView } from './views/anchoredOskView.js';
 export { default as InlinedOSKView } from './views/inlinedOskView.js';
-export { BannerController } from './banner/bannerView.js';
+export { BannerController } from './banner/bannerController.js';
 // Is referenced by at least one desktop UI module.
 export { FloatingOSKCookie as FloatingOSKViewCookie } from './views/floatingOskCookie.js';
 export { default as VisualKeyboard } from './visualKeyboard.js';
